+++ bmc-core/src/main/java/com/oracle/bmc/core/ComputeAsync.java
@@ -465,7 +465,7 @@ public interface ComputeAsync extends AutoCloseable {
                             handler);
 
     /**
-     * Performs one of the power actions (start, stop, softreset, or reset)
+     * Performs one of the power actions (start, stop, softreset, softstop, or reset)
      * on the specified instance.
      * <p>
      **start** - power on
@@ -473,6 +473,8 @@ public interface ComputeAsync extends AutoCloseable {
      **stop** - power off
      * <p>
      **softreset** - ACPI shutdown and power on
+     * <p>
+     **softstop** - signal the instance operating system to shutdown gracefully
      * <p>
      **reset** - power off and power on
      * <p>